[wptrunner] Re-enable secure WebSocket server
Commit bfef1f20a419d24633e48d24c14e6a7503e1d48c organized WPT
configuration management into a dedicated class. As part of this change,
it modified the way certain properties were initialized. The original
implementation included the following code:

    config = serve.merge_json(default_config, local_config)

The `merge_json` method operated on object values recursively. Any
properties in `default_config["ports"]` that were *not* present in
`local_config["ports"]` were persisted.

The commit referenced above translated that code to:

    config = serve.Config(override_ssl_env=self.ssl_env, **default_config)
    config.ports = {
        "http": [8000, 8001],
        "https": [8443],
        "ws": [8888]
    }

In this factoring, the `config.ports` property is completely
overwritten.  Because the `default_config` object contains a "wss"
property that is not present in the override, this new approach disables
the default secure WebSocket server.

Re-implement the original behavior by selectively overwriting only the
specified configuration properties and persisting any
previously-existing properties.