[mediasource] Automate media element playback
The Google Chrome browser has recently tightened its policy for
script-initiated media playback [1]. This interferes with tests which
assume that the element's `play` method will always cause the media to

This class of restriction is expected to be implemented by more browers,
so the testdriver.js library has been built as a generalized solution.
It allows tests to satisfy the constraints by simulating user input.

Use the testharness.js library to perform protected API calls as though
they were initiated by a user.

[1] https://developers.google.com/web/updates/2017/09/autoplay-policy-changes