[webaudio] Automate ctx creation & media playback
The Google Chrome browser has recently tightened its policy for
script-initiated audio context creation and media playback [1]. This
interferes with tests which rely on the typical behavior of those APIs.

This class of restriction is expected to be implemented by more browers,
so the testdriver.js library has been built as a generalized solution.
It allows tests to satisfy the constraints by simulating user input.

Use the testharness.js library to perform protected API calls as though
they were initiated by a user.

[1] https://developers.google.com/web/updates/2017/09/autoplay-policy-changes