[testharness.js] Honor promises from cleanup fns
Extend the implementation of the `Test#add_cleanup` method to detect
cases where the provided function returns a "thenable" object. In such
cases, defer test completion until all such eventual values settle. If
any eventual value is rejected, report a harness error (as this
represents an unpredictable state that may invalidate subsequent tests).

This behavior prevents tests from being run synchronously. In order to
promote predictability in test harness behavior (and to avoid
invalidating existing tests which rely on synchronicity), only expose
the new behavior for tests defined using the `promise_test` API (as
these tests were originally implemented to run in separate turns of the
event loop).

Introduce `Tests#set_status` in order to de-couple status setting from
remote context closing. This is necessary because uncaught errors may
require asynchronous cleanup, making it inappropriate to synchronously
close communication channels with remote contexts in such cases.

Ensure that this implementation does not introduce a new dependency on
the global `Promise` constructor in order to avoid interfering with
tests that do not require it (as doing so would invalidate those tests
in contexts where `Promise` is not available).

The return value of the function provided to `Test#add_cleanup` is only
meaningful for tests created with the `promise_test` API, and in those
cases, only when it is a "thenable" object. If a cleanup function
returns an invalid value, this may reflect a programming error that will
silently degrade test stability.

Detect cases where an invalid value is returned and trigger immediate
harness failure, helping contributors identify mistakes.

Extend test coverage to demonstrate that the new behavior functions as
expected and that these changes do not have unintended effects on
existing functionality.