[docs] Add information on virtualenv
Many of the subcommands of the WPT CLI require the `virtualenv` utility,
but the installation instructions did not instruct contributors to
install it. Add an installation step describing how to install the
dependency.

The 2.7.9 release of Python was the first to include the Pip utility
[1]. Further specify the minimum version of Python to this release in
order to limit the number of instructions.

[1] https://www.python.org/downloads/release/python-279/