PARTIAL [docs] Incorporate wptserve documentation
While Sphinx is capable of including the contents of files which exist
outside of the documentation root (via the `include` directive), it
interprets such text in the context of the including file. That makes
the mechanism inappropriate for including documentation that references
still other files using relative paths.

To address this, specify the root of the repository as the root of the
documentation build process. Specify a safe list of directories to
include. Remove the files this change obviates. Fix or ignore
infractions in the newly-incorporated files.

TODO: Update the deployment process to publish only the documentation...