[docs] Consolidate info on essential properties
As instructional content is moved from the README document to more
targeted locations within the `docs/` directory, it is becoming
increasingly important to direct readers to the documentation website.

Promote the existing content to the introduction of the README.
Introduce references to wpt.fyi and web-platform-tests.live. Add
references that introduce readers to essential technology like git and
IRC. Remove the reference to the wptserve documentation because that is
now included in the project documentation website.

Manually synchronize the README introduction and the documentation
website's introduction. (A future patch will remove this duplication
through an extension to the documentation's build process.)