[wptserve] Simplify "sub" feature
The "sub" feature implements an optional parameter which controls
whether or not HTML escaping occurs. This opt-in behavior was only used
by a single resource in WPT:

    $ git grep -E "[&?]pipe=.*sub\("
    import-maps/resources/test-helper.js:        "/import-maps/static-import.js?pipe=sub(none)&url=" +

This additional functionality complicates the feature, adding to the
amount of information users must consume in order to use it. It also
adds dissonance with the otherwise-equivalent filename-based mechanism
for enabling substitution (i.e. `test-helper.sub.js`).

Remove the functionality and re-implement the existing usage to preserve
expected behavior.