[workers] Correct variable name
This test generates subtests based on the value of a binding named
`unexpected`. Prior to this commit, no such binding was declared by the
test, so it failed to declare any subtests and instead produced a
ReferenceError. Currently, testharness.js interprets this result as a
failing "single-page test," although it will soon be more accurately
reported as a harness error [1].

Update the test to declare a binding named `unexpected`.

[1] https://github.com/web-platform-tests/rfcs/blob/master/rfcs/single_test.md