[fetch] Correct test bugs
Prior to this commit, this test referenced a non-existent binding while
generating subtest titles. This produced an uncaught exception in all
implementations. Currently, testharness.js interprets this result as a
failing "single-page test," although it will soon be more accurately
reported as a harness error [1].

Remove the unresolvable reference and improve the subtest title to more
accurately describe the behavior under test.

[1] https://github.com/web-platform-tests/rfcs/blob/master/rfcs/single_test.md