[infra] Rename test
"wptrunner" loads tests over an HTTPS connection if their filename
includes the string `.https`. This test requires HTTPS, but it does not
include that filename pattern. An independent test runner ensures that
the test is loaded over HTTPS. By renaming the test, this patch
simplifies the change necessary to refactor it for use by wptrunner.