[infra] Refactor "functional" tests
Prior to this commit, the modified tests were executed using a test
runner designed specifically for this purpose. This was suboptimal
because the dedicated code made the project infrastructure more
difficult to learn and maintain.

Refactor the tests so they can be executed using the same tooling as the
rest of the project, namely wptrunner. Relocate the tests to the
directory where other infrastructure tests have already been authored.

Update testharness.js to tolerate path names which include more than one
period (".") character.

Remove the infrastructure that was previously required to execute the
tests.

Update wptrunner to disable two optimizations (namely, the disabling of
output and the filtering of harness messages) when it executes the new
tests.