[infra] Invert variant
These tests were authored to be interpreted by a custom test runner
which removed the global Promise constructor by default. This is not the
case in wptrunner, making the semantics of a variant named
"keep-promise" somewhat confusing.

Invert the variant to "remove-promise" and remove the use of the variant
(and the corresponding utility script) from tests which are not expected
to pass in the absence of Promises. These changes were generated
programmatically using the following commands:

    $ git grep -l ?keep-promise infrastructure/idlharness.js/ | \
      xargs grep -L '=""' | \
      xargs sed -e '/="variant"/d' -e '/variant-promise/d'
    $ git grep -l ?keep-promise infrastructure/{idl,test}harness.js/ | \
      xargs sed -i 's/\?keep-promise/?remove-promise/g'

Update the test metadata accordingly.